Add concept of form groups
Likely to be renamed to form sets, this allows for several forms to be
stapled together in a declared sort order, also allowing for overridden form
titles and instructions, and to be associated with a help topic. The concept
here is to decouple the hard-coded fields used by osTicket natively and
allow the user to declare the fields required or optional for submission of
ticket information.