Initial sketch of dynamic lists
Support for creation of new lists, addition, deletion, and edits of existing
lists. Support for adding fields to forms that are drop-downs of dynamic
lists -- typeahead still to come. Support ID value in the database so that
updates to the value of a list item will be reflected in entries for that
value. Deletes for list items don't actually remove the database record,
they just drop the item from the list. The list item still exists so that be
displayed on the screen.