Add date and time field with some basic options
Allow the user to set min and max dates for the reasonable timerange for the
field. Also, allow the date/time to be kept relative to each user's clock
(stored in GMT time) or kept absolute (stored exactly as entered and not
translated). Also, permit the form to be a date field or a date and time
field.